Move workload registries to apps package
