+++ hack/.golint_failures
@@ -243,6 +243,12 @@ pkg/registry/admissionregistration/initializerconfiguration/storage
 pkg/registry/admissionregistration/mutatingwebhookconfiguration/storage
 pkg/registry/admissionregistration/rest
 pkg/registry/admissionregistration/validatingwebhookconfiguration/storage
+pkg/registry/apps/daemonset
+pkg/registry/apps/daemonset/storage
+pkg/registry/apps/deployment
+pkg/registry/apps/deployment/storage
+pkg/registry/apps/replicaset
+pkg/registry/apps/replicaset/storage
 pkg/registry/apps/rest
 pkg/registry/apps/statefulset
 pkg/registry/apps/statefulset/storage
@@ -296,14 +302,8 @@ pkg/registry/core/service/storage
 pkg/registry/core/serviceaccount/storage
 pkg/registry/events/rest
 pkg/registry/extensions/controller/storage
-pkg/registry/extensions/daemonset
-pkg/registry/extensions/daemonset/storage
-pkg/registry/extensions/deployment
-pkg/registry/extensions/deployment/storage
 pkg/registry/extensions/ingress
 pkg/registry/extensions/ingress/storage
-pkg/registry/extensions/replicaset
-pkg/registry/extensions/replicaset/storage
 pkg/registry/extensions/rest
 pkg/registry/networking/networkpolicy/storage
 pkg/registry/networking/rest
