+++ pkg/registry/BUILD
@@ -27,6 +27,9 @@ filegroup(
         "//pkg/registry/admissionregistration/rest:all-srcs",
         "//pkg/registry/admissionregistration/validatingwebhookconfiguration:all-srcs",
         "//pkg/registry/apps/controllerrevision:all-srcs",
+        "//pkg/registry/apps/daemonset:all-srcs",
+        "//pkg/registry/apps/deployment:all-srcs",
+        "//pkg/registry/apps/replicaset:all-srcs",
         "//pkg/registry/apps/rest:all-srcs",
         "//pkg/registry/apps/statefulset:all-srcs",
         "//pkg/registry/authentication/rest:all-srcs",
@@ -66,11 +69,8 @@ filegroup(
         "//pkg/registry/events/event:all-srcs",
         "//pkg/registry/events/rest:all-srcs",
         "//pkg/registry/extensions/controller/storage:all-srcs",
-        "//pkg/registry/extensions/daemonset:all-srcs",
-        "//pkg/registry/extensions/deployment:all-srcs",
         "//pkg/registry/extensions/ingress:all-srcs",
         "//pkg/registry/extensions/podsecuritypolicy:all-srcs",
-        "//pkg/registry/extensions/replicaset:all-srcs",
         "//pkg/registry/extensions/rest:all-srcs",
         "//pkg/registry/networking/networkpolicy:all-srcs",
         "//pkg/registry/networking/rest:all-srcs",
