+++ pkg/registry/apps/daemonset/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // Package daemonset provides Registry interface and its RESTStorage
 // implementation for storing DaemonSet api objects.
-package daemonset // import "k8s.io/kubernetes/pkg/registry/extensions/daemonset"
+package daemonset // import "k8s.io/kubernetes/pkg/registry/apps/daemonset"
