+++ pkg/registry/apps/daemonset/strategy.go
