+++ pkg/registry/apps/daemonset/strategy_test.go
