+++ pkg/registry/apps/deployment/doc.go
@@ -14,4 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package deployment // import "k8s.io/kubernetes/pkg/registry/extensions/deployment"
+package deployment // import "k8s.io/kubernetes/pkg/registry/apps/deployment"
