+++ pkg/registry/apps/deployment/storage/BUILD
@@ -10,7 +10,7 @@ go_test(
     name = "go_default_test",
     srcs = ["storage_test.go"],
     embed = [":go_default_library"],
-    importpath = "k8s.io/kubernetes/pkg/registry/extensions/deployment/storage",
+    importpath = "k8s.io/kubernetes/pkg/registry/apps/deployment/storage",
     deps = [
         "//pkg/apis/autoscaling:go_default_library",
         "//pkg/apis/core:go_default_library",
@@ -36,7 +36,7 @@ go_test(
 go_library(
     name = "go_default_library",
     srcs = ["storage.go"],
-    importpath = "k8s.io/kubernetes/pkg/registry/extensions/deployment/storage",
+    importpath = "k8s.io/kubernetes/pkg/registry/apps/deployment/storage",
     deps = [
         "//pkg/apis/apps/v1beta1:go_default_library",
         "//pkg/apis/apps/v1beta2:go_default_library",
@@ -46,7 +46,7 @@ go_library(
         "//pkg/apis/extensions:go_default_library",
         "//pkg/apis/extensions/v1beta1:go_default_library",
         "//pkg/apis/extensions/validation:go_default_library",
-        "//pkg/registry/extensions/deployment:go_default_library",
+        "//pkg/registry/apps/deployment:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
