+++ pkg/registry/apps/replicaset/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // Package replicaset provides Registry interface and it's RESTStorage
 // implementation for storing ReplicaSet api objects.
-package replicaset // import "k8s.io/kubernetes/pkg/registry/extensions/replicaset"
+package replicaset // import "k8s.io/kubernetes/pkg/registry/apps/replicaset"
