+++ pkg/registry/apps/replicaset/storage/storage.go
@@ -39,7 +39,7 @@ import (
 	"k8s.io/kubernetes/pkg/printers"
 	printersinternal "k8s.io/kubernetes/pkg/printers/internalversion"
 	printerstorage "k8s.io/kubernetes/pkg/printers/storage"
-	"k8s.io/kubernetes/pkg/registry/extensions/replicaset"
+	"k8s.io/kubernetes/pkg/registry/apps/replicaset"
 )
 
 // ReplicaSetStorage includes dummy storage for ReplicaSets and for Scale subresource.
