+++ pkg/registry/apps/replicaset/strategy_test.go
