+++ pkg/registry/apps/rest/BUILD
@@ -13,10 +13,10 @@ go_library(
         "//pkg/api/legacyscheme:go_default_library",
         "//pkg/apis/apps:go_default_library",
         "//pkg/registry/apps/controllerrevision/storage:go_default_library",
+        "//pkg/registry/apps/daemonset/storage:go_default_library",
+        "//pkg/registry/apps/deployment/storage:go_default_library",
+        "//pkg/registry/apps/replicaset/storage:go_default_library",
         "//pkg/registry/apps/statefulset/storage:go_default_library",
-        "//pkg/registry/extensions/daemonset/storage:go_default_library",
-        "//pkg/registry/extensions/deployment/storage:go_default_library",
-        "//pkg/registry/extensions/replicaset/storage:go_default_library",
         "//vendor/k8s.io/api/apps/v1:go_default_library",
         "//vendor/k8s.io/api/apps/v1beta1:go_default_library",
         "//vendor/k8s.io/api/apps/v1beta2:go_default_library",
