+++ pkg/registry/extensions/rest/BUILD
@@ -12,12 +12,12 @@ go_library(
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
         "//pkg/apis/extensions:go_default_library",
+        "//pkg/registry/apps/daemonset/storage:go_default_library",
+        "//pkg/registry/apps/deployment/storage:go_default_library",
+        "//pkg/registry/apps/replicaset/storage:go_default_library",
         "//pkg/registry/extensions/controller/storage:go_default_library",
-        "//pkg/registry/extensions/daemonset/storage:go_default_library",
-        "//pkg/registry/extensions/deployment/storage:go_default_library",
         "//pkg/registry/extensions/ingress/storage:go_default_library",
         "//pkg/registry/extensions/podsecuritypolicy/storage:go_default_library",
-        "//pkg/registry/extensions/replicaset/storage:go_default_library",
         "//pkg/registry/networking/networkpolicy/storage:go_default_library",
         "//vendor/k8s.io/api/extensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/generic:go_default_library",
